# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIKey:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      key:
        description: Key is the plaintext key, set only in the response that creates
          it.
        type: string
      label:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
  models.ChangePasswordResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
...
  models.ContactRequest:
    properties:
      name:
        example: Jane Doe
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      label:
        type: string
    type: object
  models.CreateAPIKeyResponse:
    properties:
      api_key:
        $ref: '#/definitions/models.APIKey'
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
host: localhost:5174
info:
  contact: {}
  description: REST API for sending and receiving SMS messages via a USB GSM modem.
  license:
    name: GPL-3.0
  title: SMS Gateway API
  version: "1.0"
paths:
  /api/v1/apikeys:
    get:
      description: Returns API keys for the authenticated user, newest first. Admins
        see all keys.
      parameters:
      - description: Maximum keys to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Keys to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total matching keys is returned in the X-Total-Count header
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Generates a new API key for the authenticated user with an optional
        label.
      parameters:
      - description: API key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - API Keys
  /api/v1/apikeys/{id}:
    delete:
      description: Deactivates an API key by its unique identifier. Non-admins can
        only deactivate their own keys.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: API key deactivated'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate API key
      tags:
      - API Keys
  /api/v1/apikeys/{id}/delete:
    delete:
      description: Permanently deletes an API key by its unique identifier. Non-admins
        can only delete their own keys.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: API key deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - API Keys
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Allows an authenticated user to change their password by providing
        the current and new passwords. Every existing token is revoked; the response
        carries a new one for the caller.
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChangePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with username and password, returning a JWT
        token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Revokes every token issued to the authenticated user, logging them
        out on all devices.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: logged out'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Auth
  /api/v1/contacts:
    delete:
      description: Removes the saved name for a phone number. Messages are not affected.
      parameters:
      - description: Phone number of the contact
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete contact name
      tags:
      - Contacts
    get:
      description: Returns saved contact names ordered by name.
      parameters:
      - description: Only contacts whose name or number contains this text
        in: query
        name: q
        type: string
      - description: Maximum contacts to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Contacts to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total matching contacts is returned in the X-Total-Count header
          schema:
            items:
              $ref: '#/definitions/models.Contact'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List contacts
      tags:
      - Contacts
    put:
      consumes:
      - application/json
      description: Sets the display name for a phone number, creating the contact
        if needed.
      parameters:
      - description: Phone number of the contact
        in: query
        name: phone
        required: true
        type: string
      - description: Contact name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Save contact name
      tags:
      - Contacts
...
  /api/v1/modem/at:
    post:
      consumes:
      - application/json
      description: Sends a raw AT command to the GSM modem and returns the response.
        Requires admin privileges.
      parameters:
      - description: AT command to send
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ATCommandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ATCommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Dangerous or unrecognised command sent without confirm
          schema:
            $ref: '#/definitions/models.ATConfirmationRequired'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send AT command
      tags:
      - Modem
  /api/v1/modem/at/commands:
    get:
      description: Returns documentation for common AT commands, with the risk of
        running each, and the command names the modem reports through AT+CLAC (null
        when unsupported). Requires admin privileges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ATCatalogResponse'
      security:
      - BearerAuth: []
      summary: AT command reference
      tags:
      - Modem
  /api/v1/modem/info:
    get:
      description: Returns the line provider and network, the SIM phone number, ICCID
        and IMSI, and the modem IMEI, manufacturer, model and firmware. A field is
        empty when the modem or SIM cannot report it; many SIMs do not store their
        phone number. The details are read once and cached until the service restarts;
        pass refresh=true to read them from the modem again. Requires admin privileges.
      parameters:
      - description: Read the details from the modem again instead of using the cached
          ones
        in: query
        name: refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get modem and SIM details
      tags:
      - Modem
  /api/v1/modem/profiles:
    get:
      description: Returns the carrier setup profiles, ordered by name. Requires admin
        privileges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ModemProfile'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List modem profiles
      tags:
      - Modem
    post:
      consumes:
      - application/json
      description: 'Creates a carrier setup profile: ordered steps of AT commands,
        each with an optional case-insensitive regular expression its response must
        match. Requires admin privileges.'
      parameters:
      - description: Profile to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ModemProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ModemProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create modem profile
      tags:
      - Modem
  /api/v1/modem/profiles/{id}:
    delete:
      description: Permanently deletes a carrier setup profile. Requires admin privileges.
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: modem profile deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete modem profile
      tags:
      - Modem
    get:
      description: Returns a carrier setup profile by ID. Requires admin privileges.
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get modem profile
      tags:
      - Modem
    put:
      consumes:
      - application/json
      description: Replaces a carrier setup profile's name, description, notes and
        steps. Requires admin privileges.
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: New profile contents
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ModemProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update modem profile
      tags:
      - Modem
  /api/v1/modem/signal:
    get:
      description: Returns the GSM modem signal strength and quality assessment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemSignalResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get modem signal strength
      tags:
      - Modem
  /api/v1/modem/status:
    get:
      description: Checks whether the GSM modem is responsive and returns its current
        status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemStatusResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ModemStatusResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get modem status
      tags:
      - Modem
  /api/v1/sms/{id}:
    delete:
      description: Deletes a single SMS message by its unique identifier.
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete message
      tags:
      - SMS
    get:
      description: Returns a single SMS message by its unique identifier.
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get message by ID
      tags:
      - SMS
  /api/v1/sms/{id}/read:
    put:
      description: Marks an inbound message as read, changing its status from "received"
        to "read".
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Mark message as read
      tags:
      - SMS
  /api/v1/sms/{id}/unread:
    put:
      description: Marks an inbound message as unread, changing its status from "read"
        back to "received".
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Mark message as unread
      tags:
      - SMS
  /api/v1/sms/conversations:
    delete:
      description: Deletes every inbound and outbound message for the phone number.
      parameters:
      - description: Phone number of the conversation
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConversationUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete conversation
      tags:
      - SMS
    get:
      description: Returns inbound and outbound messages grouped by phone number,
        most recent activity first. Each entry carries the latest message and message/unread
        counts.
      parameters:
      - description: Only conversations with a message whose number or body contains
          this text
        in: query
        name: q
        type: string
      - description: Maximum conversations to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Conversations to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total matching conversations is returned in the X-Total-Count
            header
          schema:
            items:
              $ref: '#/definitions/models.Conversation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List conversations
      tags:
      - SMS
...
  /api/v1/sms/stats:
    get:
      description: Returns message counts aggregated by direction and status. Lets
        clients show accurate totals without downloading every message.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageStats'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get message statistics
      tags:
      - SMS
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix (e.g. "Bearer eyJhbG...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"